requirements.txt
click==8.2.1
greenlet==3.2.3
psycopg2-binary==2.9.10
python-dotenv==1.1.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
